requirements.txt
openai
pandas
openpyxl
python-dotenv
pydantic
loguru
faiss-cpu
sentence-transformers
pytest
streamlit
